25 animated transport (#34)
* added performance logging and faster city gen

* first stage of animated transport

* car rendering now working

- cars in different colors
- cars are renderd with the right depth
- movement speed is dependend on game speed

* updated yaml file to current version

* fix warnings and errors with msvc

* fixed release.yml

* fixed some path finding problems and daily updates

* fixed stopping threads while closing the game

+++ src/world/loan.cpp
@@ -42,13 +42,13 @@ namespace world
         utils::time::Date installmentDate = m_startDay;
         if (m_startDay.getDay() > 1)
         {
-            size_t lastDayOfMonth = utils::time::lastDayOfMonth(m_startDay.getYear(), m_startDay.getMonth());
+            uint16_t lastDayOfMonth = utils::time::lastDayOfMonth(m_startDay.getYear(), m_startDay.getMonth());
             installmentDate.addDay(lastDayOfMonth - m_startDay.getDay());
         }
         double repaymentAmount = m_amount;
         for (size_t i = 0; i < numberOfInstallments; ++i)
         {
-            size_t lastDayOfMonth = utils::time::lastDayOfMonth(installmentDate.getYear(), installmentDate.getMonth());
+            uint16_t lastDayOfMonth = utils::time::lastDayOfMonth(installmentDate.getYear(), installmentDate.getMonth());
             installmentDate.addDay(lastDayOfMonth);
             double installmentAmount = std::round(m_amount / double(numberOfInstallments) * 100.0) / 100.0;
             repaymentAmount -= installmentAmount;
